down-scope cns client interface and move to consumer packages
