+++ cns/singletenantcontroller/kubecontroller/crdreconciler.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	"github.com/Azure/azure-container-networking/cns"
-	cnsclient "github.com/Azure/azure-container-networking/cns/client"
 	"github.com/Azure/azure-container-networking/cns/logger"
 	"github.com/Azure/azure-container-networking/crd/nodenetworkconfig/api/v1alpha"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -17,7 +16,7 @@ import (
 type CrdReconciler struct {
 	KubeClient      KubeClient
 	NodeName        string
-	CNSClient       cnsclient.APIClient
+	CNSClient       cnsclient
 	IPAMPoolMonitor cns.IPAMPoolMonitor
 }
